Refactor dead code elimination tests for while statements
- Updated BUILD file to correct the path for lit test files.
- Simplified the passes in the README and run generated files by removing unnecessary passes.
- Adjusted indentation in input.js for better readability.
- Cleaned up output.generated.txt by removing excessive comments and irrelevant sections.
