+++ maldoca/js/ir/transforms/dead_code_elimination/tests/after_dynconstprop/README.generated.md
@@ -1,0 +1,7 @@
+To run manually:
+
+```shell
+bazel run //maldoca/js/ir:jsir_gen -- \
+  --input_file $(pwd)/maldoca/js/ir/transforms/dead_code_elimination/tests/high/input.js \
+  --passes "source2ast,ast2hir,dead_code_elimination,hir2ast,ast2source"
+```
