+++ maldoca/js/ir/transforms/dead_code_elimination/tests/after_dynconstprop/input.js
@@ -1,0 +1,919 @@
+function calculatePrice(_0x5b037d, _0x4b126e) {
+  function _0x3a3b81(_0x49082e, _0x572c13, _0x32ce12, _0x4a17d8, _0x502d0d) {
+    return _0x5ec0(_0x502d0d - -170, _0x32ce12);
+  }
+  const _0x23fb6 = {
+      'FZwIH': function (_0x3f8fa4, _0x58105b) {
+        return _0x3f8fa4(_0x58105b);
+      },
+      'IWoxy': function (_0x5d9aa6, _0x144983) {
+        return _0x5d9aa6 + _0x144983;
+      },
+      'tUwMG': "return (function() ",
+      'oJhKn': "{}.constructor(\"return this\")( )",
+      'txgFU': function (_0x30fd9f) {
+        return _0x30fd9f();
+      },
+      'nhPut': function (_0xd62405, _0x3a582a) {
+        return _0xd62405 === _0x3a582a;
+      },
+      'voRwb': "AFNky",
+      'GbKwb': "OVvjH",
+      'HeGGB': function (_0x1b7b9c, _0x4d86b2) {
+        return _0x1b7b9c !== _0x4d86b2;
+      },
+      'wZVud': "TWJvb",
+      'gIUsJ': "cQQgA",
+      'zzAKA': function (_0x508d54, _0x7ace76) {
+        return _0x508d54 === _0x7ace76;
+      },
+      'lxcov': "NqUzh",
+      'rELsr': "zLyeH",
+      'hnIaJ': function (_0x3c8875, _0xf87df5) {
+        return _0x3c8875(_0xf87df5);
+      },
+      'lLGBN': function (_0x58b6fa, _0x3d96f1) {
+        return _0x58b6fa + _0x3d96f1;
+      },
+      'lKFSI': function (_0xae036c, _0x3da58e) {
+        return _0xae036c + _0x3da58e;
+      },
+      'iMnmu': "PRoPb",
+      'hlsZB': "hoxlU",
+      'MwNVS': "debu",
+      'oMPbC': "gger",
+      'UCoqb': "stateObject",
+      'pWtFV': function (_0x584d7a, _0x301f01) {
+        return _0x584d7a(_0x301f01);
+      },
+      'VSLQK': "function *\\( *\\)",
+      'TWYxn': "\\+\\+ *(?:[a-zA-Z_$][0-9a-zA-Z_$]*)",
+      'RfJBn': function (_0x350326, _0x3ce5bd) {
+        return _0x350326(_0x3ce5bd);
+      },
+      'YoyTe': "init",
+      'GNdZT': "chain",
+      'FKoAj': "input",
+      'BOPqU': function (_0x9ffe79, _0x24bad5) {
+        return _0x9ffe79(_0x24bad5);
+      },
+      'IQePn': function (_0x554134) {
+        return _0x554134();
+      },
+      'XSQbQ': function (_0x5ab606, _0x2c62f5, _0x521648) {
+        return _0x5ab606(_0x2c62f5, _0x521648);
+      },
+      'OVlMy': "LAaXM",
+      'xyBvI': "CqrWo",
+      'NMvvT': function (_0x4b9f09, _0x2edc9e) {
+        return _0x4b9f09 + _0x2edc9e;
+      },
+      'NDoCJ': function (_0x15c537, _0x5ccac7) {
+        return _0x15c537 + _0x5ccac7;
+      },
+      'LeAqo': function (_0x1a9fb9, _0x4748d5) {
+        return _0x1a9fb9 !== _0x4748d5;
+      },
+      'lJGYk': "kbGEP",
+      'ZxHsI': function (_0x186b48, _0x429652) {
+        return _0x186b48(_0x429652);
+      },
+      'AAjwv': function (_0xc36c35, _0x5b1405) {
+        return _0xc36c35 === _0x5b1405;
+      },
+      'GFISC': "umQMd",
+      'biCsT': function (_0x227d3e) {
+        return _0x227d3e();
+      },
+      'xdsvJ': "eckNK",
+      'GGULq': "eiCBe",
+      'IWcYV': "while (true) {}",
+      'dPdUr': "counter",
+      'xDMyq': function (_0x4e7cf0, _0x5d02c3) {
+        return _0x4e7cf0 !== _0x5d02c3;
+      },
+      'yAGxJ': "vAzJS",
+      'jsIgl': "gIPTp",
+      'Tqufk': "sHFfh",
+      'UHIZG': function (_0x3459a7, _0x509008) {
+        return _0x3459a7 === _0x509008;
+      },
+      'Stjop': "bEHUp",
+      'hPHhO': "aPxIb",
+      'puyhR': "XBYMu",
+      'SRFQx': function (_0x494b19, _0x3379df) {
+        return _0x494b19 !== _0x3379df;
+      },
+      'PCfRO': "TLjgh",
+      'LMwVs': "MekFX",
+      'TDTuJ': "tlWRU",
+      'SWMyF': function (_0x1fecf2, _0x1e294e) {
+        return _0x1fecf2(_0x1e294e);
+      },
+      'TObTc': "cRXBN",
+      'tyXku': "action",
+      'ixHSk': "UuWki",
+      'ThYsS': function (_0x35c982) {
+        return _0x35c982();
+      },
+      'DSNDX': "log",
+      'WaCZA': "warn",
+      'ytswv': "info",
+      'kBqLi': "error",
+      'VCkuy': "exception",
+      'qMUag': "table",
+      'RCHoV': "trace",
+      'AzhjX': function (_0x29fbaa, _0x549bc5) {
+        return _0x29fbaa < _0x549bc5;
+      },
+      'cRwGw': "Pywil",
+      'pUMUj': "ujGre",
+      'mOHAK': function (_0x3174ca, _0x22ea39, _0x4a760c) {
+        return _0x3174ca(_0x22ea39, _0x4a760c);
+      },
+      'xfELl': function (_0x99593c, _0x5884e7) {
+        return _0x99593c * _0x5884e7;
+      },
+      'UFmYJ': function (_0x103392, _0x5e42fe) {
+        return _0x103392 > _0x5e42fe;
+      },
+      'oUcea': function (_0x3f98e0, _0x4ebf60) {
+        return _0x3f98e0 !== _0x4ebf60;
+      },
+      'nSyDi': "JsOcE",
+      'xApGy': function (_0x277418, _0x56a6fc) {
+        return _0x277418 - _0x56a6fc;
+      }
+    },
+    _0x54e352 = (function () {
+      function _0x5d0c51(_0x59f630, _0x121cbc, _0x59395a, _0x3cc6fe, _0x45e319) {
+        return _0x22ca9b(_0x59f630 - 221, _0x45e319, _0x121cbc - 1092, _0x3cc6fe - 123, _0x45e319 - 27);
+      }
+      function _0x3e073c(_0x2516c4, _0x207ef5, _0x420546, _0x2f3381, _0x452168) {
+        return _0x2d5cb7(_0x2516c4 - 366, _0x207ef5 - 47, _0x2f3381 - -952, _0x2f3381 - 435, _0x2516c4);
+      }
+      function _0x4b3646(_0x1a2583, _0x23fea6, _0xd4fa83, _0x2fbf39, _0x114a51) {
+        return _0x3a3b81(_0x1a2583 - 158, _0x23fea6 - 131, _0x23fea6, _0x2fbf39 - 145, _0x2fbf39 - 685);
+      }
+      function _0x3e58c1(_0x37f76b, _0x294670, _0x27b76b, _0x2f328d, _0x3f40f6) {
+        return _0x22ca9b(_0x37f76b - 14, _0x3f40f6, _0x27b76b - 91, _0x2f328d - 119, _0x3f40f6 - 142);
+      }
+      function _0x5765f7(_0x39f077, _0x3b6046, _0x4365f5, _0x3641b6, _0x3624bd) {
+        return _0x22ca9b(_0x39f077 - 291, _0x3641b6, _0x3624bd - 262, _0x3641b6 - 287, _0x3624bd - 219);
+      }
+      if (true) {
+        let _0x5ea015 = !![];
+        return function (_0x447810, _0x17ffae) {
+          const _0x3be97e = {
+            'GzDIq': function (_0x319bf7, _0xd41c35) {
+              function _0x5078f7(_0x5f45d4, _0x5b9a10, _0x5de014, _0x4e3cf4, _0xcc338d) {
+                return _0x5ec0(_0xcc338d - 742, _0x5f45d4);
+              }
+              return _0x319bf7(_0xd41c35);
+            },
+            'GtCQm': function (_0x535928, _0x547e85) {
+              function _0x5219e3(_0x30098e, _0x4e1b34, _0x504a6c, _0x1cbaaf, _0x1ab293) {
+                return _0x5ec0(_0x4e1b34 - 177, _0x1ab293);
+              }
+              return _0x23fb6["IWoxy"](_0x535928, _0x547e85);
+            },
+            'JdWiA': "return (function() ",
+            'JSNoO': "{}.constructor(\"return this\")( )",
+            'doDtr': function (_0x61ccf6) {
+              function _0xc4a3cd(_0x1ebe3b, _0xee6551, _0x4c032a, _0x346e46, _0x34c822) {
+                return _0x3837ef(_0x346e46 - 984, _0xee6551 - 379, _0x4c032a - 365, _0x346e46 - 236, _0x4c032a);
+              }
+              return _0x61ccf6();
+            },
+            'JaATJ': function (_0xa8b608, _0x596d22) {
+              function _0x3b8494(_0x476fe1, _0x41831f, _0x2137f1, _0x2bd392, _0x57fb17) {
+                return _0x3837ef(_0x476fe1 - 1711, _0x41831f - 313, _0x2137f1 - 492, _0x2bd392 - 90, _0x41831f);
+              }
+              return _0x23fb6["nhPut"](_0xa8b608, _0x596d22);
+            },
+            'LdRVX': "AFNky",
+            'QmJGn': "OVvjH",
+            'jFTYa': function (_0x441824, _0x4a42f4) {
+              function _0x24e1b6(_0x20927f, _0x4ddc10, _0x345e9e, _0x4c8572, _0x2c9a5d) {
+                return _0x3837ef(_0x345e9e - 975, _0x4ddc10 - 297, _0x345e9e - 491, _0x4c8572 - 16, _0x2c9a5d);
+              }
+              return _0x23fb6["HeGGB"](_0x441824, _0x4a42f4);
+            },
+            'LVHVn': "TWJvb",
+            'jZxAe': "cQQgA"
+          };
+          function _0x4f512e(_0x4bd4a7, _0x3fdcab, _0x261d49, _0x3b3e1f, _0x3d3975) {
+            return _0x4b3646(_0x4bd4a7 - 472, _0x4bd4a7, _0x261d49 - 24, _0x3d3975 - -681, _0x3d3975 - 493);
+          }
+          function _0x3837ef(_0x379d44, _0xd6ece3, _0x430353, _0x48fe3d, _0x44ac9a) {
+            return _0x5765f7(_0x379d44 - 457, _0xd6ece3 - 13, _0x430353 - 227, _0x44ac9a, _0x379d44 - -87);
+          }
+          function _0x5d9af5(_0x1ffd44, _0x1506be, _0x58f351, _0x1f9066, _0x3617d2) {
+            return _0x3e073c(_0x1ffd44, _0x1506be - 77, _0x58f351 - 176, _0x1506be - 1232, _0x3617d2 - 402);
+          }
+          function _0x11d9ec(_0x349855, _0x5734ad, _0x5b4b23, _0x715491, _0x2fb97f) {
+            return _0x3e073c(_0x2fb97f, _0x5734ad - 85, _0x5b4b23 - 267, _0x5734ad - 1099, _0x2fb97f - 190);
+          }
+          function _0x106ae2(_0x5b078c, _0x59a6d1, _0x149a11, _0x2c2aaf, _0x5aaeaf) {
+            return _0x3e073c(_0x5aaeaf, _0x59a6d1 - 18, _0x149a11 - 394, _0x2c2aaf - 1376, _0x5aaeaf - 56);
+          }
+          if (false) _0x20d949 = _0x4a0256;else {
+            const _0x3e1e80 = _0x5ea015 ? function () {
+              function _0x4da531(_0x1d1044, _0x1084a2, _0x193fb3, _0x9f01c5, _0x209212) {
+                return _0x5d9af5(_0x1084a2, _0x209212 - 367, _0x193fb3 - 109, _0x9f01c5 - 241, _0x209212 - 282);
+              }
+              function _0x1748e1(_0x347e8c, _0x4220c1, _0x3e81cf, _0x2d3393, _0x3caece) {
+                return _0x5d9af5(_0x3caece, _0x347e8c - 403, _0x3e81cf - 471, _0x2d3393 - 276, _0x3caece - 15);
+              }
+              function _0x3b85cd(_0x38f856, _0x302f8a, _0x27cf52, _0x262b31, _0x7e66a9) {
+                return _0x11d9ec(_0x38f856 - 373, _0x38f856 - -968, _0x27cf52 - 299, _0x262b31 - 362, _0x302f8a);
+              }
+              function _0x3731d4(_0x5bf3e4, _0x49ea54, _0x123074, _0x5a3d9c, _0xc70fc0) {
+                return _0x106ae2(_0x5bf3e4 - 259, _0x49ea54 - 283, _0x123074 - 445, _0x49ea54 - -128, _0xc70fc0);
+              }
+              function _0x452a1a(_0x20cd1e, _0x4f7756, _0x2f9c33, _0x2f0918, _0x97bb3e) {
+                return _0x106ae2(_0x20cd1e - 461, _0x4f7756 - 187, _0x2f9c33 - 91, _0x4f7756 - -1360, _0x2f0918);
+              }
+              if (_0x3be97e["JaATJ"]("AFNky", "OVvjH")) {
+                if (_0x268116) {
+                  const _0x1ea2b4 = _0x1268d3["apply"](_0x2c200b, arguments);
+                  return _0x370a70 = null, _0x1ea2b4;
+                }
+              } else {
+                if (_0x17ffae) {
+                  if (_0x3be97e["jFTYa"]("TWJvb", "cQQgA")) {
+                    const _0x28b6a5 = _0x17ffae["apply"](_0x447810, arguments);
+                    return _0x17ffae = null, _0x28b6a5;
+                  } else {
+                    const _0x359357 = tpCaPn["GzDIq"](_0x1ccbb6, tpCaPn["GtCQm"](tpCaPn["GtCQm"](tpCaPn["JdWiA"], tpCaPn["JSNoO"]), ");"));
+                    _0x2c9c20 = tpCaPn["doDtr"](_0x359357);
+                  }
+                }
+              }
+            } : function () {};
+            return _0x5ea015 = ![], _0x3e1e80;
+          }
+        };
+      } else _0xc4a413 = bazvTn[_0x4b3646(0x3d4, 'u0Xg', 0x28e, 0x362, 0x3e0)](_0x5cf055, bazvTn[_0x3e58c1(-0x1a9, -0x1e1, -0x1cb, -0x23e, '$xQO')](bazvTn[_0x5d0c51(0x217, 0x14e, 0xe5, 0x14e, 't!4C')](bazvTn[_0x3e58c1(-0x19a, -0x170, -0x13e, -0x198, 'f6aq')], bazvTn[_0x5765f7(-0xfa, -0x22d, -0x129, '2@7B', -0x183)]), ');'))();
+    }());
+  (function () {
+    function _0x25a218(_0x2ce3a1, _0x1cefa3, _0x446213, _0x402655, _0x430a72) {
+      return _0x22ca9b(_0x2ce3a1 - 71, _0x402655, _0x1cefa3 - 619, _0x402655 - 465, _0x430a72 - 104);
+    }
+    function _0x51f4e0(_0x228773, _0xbf116c, _0xcbdb37, _0x36d60b, _0x9b28bd) {
+      return _0x22ca9b(_0x228773 - 204, _0x228773, _0x9b28bd - 1662, _0x36d60b - 202, _0x9b28bd - 420);
+    }
+    function _0x33a3c3(_0x10ba9c, _0x28095f, _0x286159, _0x34aaae, _0x45f54b) {
+      return _0x22ca9b(_0x10ba9c - 55, _0x28095f, _0x45f54b - 69, _0x34aaae - 208, _0x45f54b - 118);
+    }
+    function _0x3a17e4(_0x4a9453, _0x4e8d6e, _0xc2b4f0, _0x568f01, _0x149233) {
+      return _0x1e1698(_0x568f01 - -160, _0x4e8d6e - 355, _0xc2b4f0 - 1, _0x568f01 - 469, _0xc2b4f0);
+    }
+    function _0x55d0ca(_0x4e2841, _0x1ea0f7, _0x3828a6, _0x295fe6, _0x3a1309) {
+      return _0x1e1698(_0x4e2841 - 1043, _0x1ea0f7 - 80, _0x3828a6 - 370, _0x295fe6 - 426, _0x295fe6);
+    }
+    false ? bazvTn[_0x25a218(-0x78, -0x59, 0x28, 'pOH8', -0x33)](_0x33f6f2) : _0x54e352(this, function () {
+      function _0x1e7a6f(_0x225f2a, _0x4b678d, _0x4aec0b, _0x188244, _0x36fb0c) {
+        return _0x55d0ca(_0x4b678d - -3, _0x4b678d - 412, _0x4aec0b - 101, _0x4aec0b, _0x36fb0c - 386);
+      }
+      function _0x4c2b48(_0x5ae9d7, _0x39db7a, _0x47754b, _0x11f67b, _0x447dce) {
+        return _0x25a218(_0x5ae9d7 - 154, _0x47754b - 813, _0x47754b - 308, _0x39db7a, _0x447dce - 260);
+      }
+      const _0x4a4fcb = {
+        'MfXlV': function (_0x9334ca, _0x1863fb) {
+          function _0x12d6e0(_0xeee6a0, _0x16f120, _0x14ddf2, _0x4f386d, _0x5e295d) {
+            return _0x5ec0(_0x4f386d - 800, _0x14ddf2);
+          }
+          return _0x23fb6["IWoxy"](_0x9334ca, _0x1863fb);
+        },
+        'yRBXX': "debu",
+        'gCxBj': "gger",
+        'GPttG': "stateObject",
+        'COGkj': function (_0x388be9, _0x81a4c2) {
+          function _0x13d50c(_0x40c154, _0x2863bd, _0x4c4835, _0x591071, _0x46db58) {
+            return _0x4c2b48(_0x40c154 - 358, _0x2863bd, _0x40c154 - -601, _0x591071 - 335, _0x46db58 - 449);
+          }
+          return _0x388be9(_0x81a4c2);
+        },
+        'djEKp': "function *\\( *\\)",
+        'wloMG': "\\+\\+ *(?:[a-zA-Z_$][0-9a-zA-Z_$]*)",
+        'rgryI': function (_0x3a859a, _0x59d151) {
+          function _0x2d6be1(_0x5d918c, _0x120115, _0x16cf8f, _0x17d454, _0x5e0407) {
+            return _0x18dbc0(_0x120115, _0x120115 - 192, _0x16cf8f - 499, _0x17d454 - 471, _0x5e0407 - 173);
+          }
+          return _0x3a859a(_0x59d151);
+        },
+        'lBFyN': "init",
+        'BEUkj': "chain",
+        'EYYLF': "input",
+        'zzuOM': function (_0x19e771, _0x201ffb) {
+          function _0x26a7d2(_0x427817, _0xd1ee8f, _0x58e16a, _0x1ba376, _0xcb5e93) {
+            return _0x315929(_0x427817 - 322, _0xd1ee8f - 278, _0x58e16a - 265, _0x427817, _0xd1ee8f - 472);
+          }
+          return _0x19e771(_0x201ffb);
+        },
+        'pFAKE': function (_0x38383a) {
+          function _0x1c46eb(_0x5af4d8, _0x2add03, _0x3ebda5, _0x10d887, _0x4d3fc0) {
+            return _0x4c2b48(_0x5af4d8 - 6, _0x5af4d8, _0x4d3fc0 - -181, _0x10d887 - 155, _0x4d3fc0 - 295);
+          }
+          return _0x38383a();
+        },
+        'usCQH': function (_0x5bb395, _0x3fd96e, _0x4442ac) {
+          function _0x33df85(_0x1da37a, _0x359f30, _0x5cbefe, _0x42c2e6, _0x151242) {
+            return _0x4c2b48(_0x1da37a - 96, _0x359f30, _0x42c2e6 - -265, _0x42c2e6 - 118, _0x151242 - 237);
+          }
+          return _0x5bb395(_0x3fd96e, _0x4442ac);
+        }
+      };
+      function _0x29c487(_0x1775a0, _0x2b00ee, _0x5c8314, _0x248b00, _0x2d49c4) {
+        return _0x3a17e4(_0x1775a0 - 41, _0x2b00ee - 256, _0x5c8314, _0x2b00ee - 1054, _0x2d49c4 - 458);
+      }
+      function _0x18dbc0(_0x5e0719, _0x2acec8, _0x334f5f, _0x55b9d4, _0x3db894) {
+        return _0x51f4e0(_0x5e0719, _0x2acec8 - 468, _0x334f5f - 8, _0x55b9d4 - 238, _0x3db894 - -1530);
+      }
+      function _0x315929(_0x949509, _0x4413ef, _0x5bb718, _0x2a874f, _0x886e6b) {
+        return _0x55d0ca(_0x886e6b - -906, _0x4413ef - 288, _0x5bb718 - 56, _0x2a874f, _0x886e6b - 218);
+      }
+      if (false) (function () {
+        return ![];
+      }[_0x4c2b48(0x3c7, 'qNNK', 0x3f5, 0x43e, 0x384) + _0x315929(0x8, 0x122, 0xf0, 'QqQ*', 0xbf) + 'r'](EEyZnd[_0x29c487(0x359, 0x33b, 'nU#o', 0x2b2, 0x2b8)](EEyZnd[_0x4c2b48(0x319, 'PjKF', 0x2ad, 0x20e, 0x30d)], EEyZnd[_0x4c2b48(0x36c, 'f6aq', 0x301, 0x2d4, 0x354)]))[_0x315929(0x220, 0x246, 0xe3, 'RVOB', 0x19f)](EEyZnd[_0x315929(0x10f, 0x1a8, 0x1dc, '@fQ!', 0x1c6)]));else {
+        const _0x5a0c95 = new RegExp("function *\\( *\\)"),
+          _0x4a0f5e = new RegExp("\\+\\+ *(?:[a-zA-Z_$][0-9a-zA-Z_$]*)", "i"),
+          _0x16c109 = _0x79c5f6("init");
+        !_0x5a0c95["test"](_0x23fb6["NMvvT"](_0x16c109, "chain")) || !_0x4a0f5e["test"](_0x23fb6["NDoCJ"](_0x16c109, "input")) ? false ? EEyZnd[_0x315929(0x1c2, 0xd2, 0xbe, '7FpC', 0xed)](_0x17784f, -0x117e * 0x1 + -0x293 * -0x7 + -0x1b * 0x5) : _0x16c109("0") : true ? _0x79c5f6() : EEyZnd[_0x315929(0x1a8, 0xaa, 0x1ac, 'bh#1', 0x177)](_0x2826ba, this, function () {
+          function _0x3a6993(_0x564139, _0x2e9a1a, _0x4990, _0x21e173, _0x4d4a0f) {
+            return _0x315929(_0x564139 - 336, _0x2e9a1a - 263, _0x4990 - 112, _0x2e9a1a, _0x4d4a0f - -658);
+          }
+          const _0x16b499 = new _0x223f50(EEyZnd["djEKp"]),
+            _0x270d77 = new _0x5a1f26(EEyZnd["wloMG"], "i");
+          function _0x2d5728(_0x518154, _0x470a62, _0x3d4904, _0x3630d4, _0x19b8de) {
+            return _0x1e7a6f(_0x518154 - 28, _0x3d4904 - 173, _0x518154, _0x3630d4 - 210, _0x19b8de - 23);
+          }
+          function _0x42f526(_0x353144, _0x15b6eb, _0x1db1aa, _0xb28b53, _0x2cf9c5) {
+            return _0x1e7a6f(_0x353144 - 88, _0x1db1aa - -32, _0x2cf9c5, _0xb28b53 - 491, _0x2cf9c5 - 274);
+          }
+          function _0x3a5c64(_0x890f3d, _0x3fba11, _0x4ad7f6, _0x5418c5, _0xf5ab03) {
+            return _0x1e7a6f(_0x890f3d - 367, _0x4ad7f6 - 42, _0x5418c5, _0x5418c5 - 179, _0xf5ab03 - 445);
+          }
+          const _0x16bd89 = EEyZnd["rgryI"](_0x4d7c19, EEyZnd["lBFyN"]);
+          function _0x2936d4(_0x4f42bb, _0x1ca2ba, _0x537e84, _0x735b1c, _0x4a07bd) {
+            return _0x4c2b48(_0x4f42bb - 404, _0x4a07bd, _0x1ca2ba - -222, _0x735b1c - 163, _0x4a07bd - 495);
+          }
+          !_0x16b499["test"](EEyZnd["MfXlV"](_0x16bd89, EEyZnd["BEUkj"])) || !_0x270d77["test"](EEyZnd["MfXlV"](_0x16bd89, EEyZnd["EYYLF"])) ? EEyZnd["zzuOM"](_0x16bd89, "0") : EEyZnd["pFAKE"](_0x9e7260);
+        })();
+      }
+    })();
+  }());
+  const _0x2fad7f = (function () {
+      const _0x5851e9 = {
+        'uUSRr': function (_0x564f88, _0x176b0d) {
+          function _0x5c6314(_0x25a523, _0x4088db, _0x3935af, _0x1ea420, _0x49ece7) {
+            return _0x5ec0(_0x25a523 - -28, _0x49ece7);
+          }
+          return _0x564f88(_0x176b0d);
+        },
+        'AWARS': "while (true) {}",
+        'oESVL': "counter",
+        'BjFNt': function (_0x10a38e, _0x635b4b) {
+          function _0x261e2b(_0x50103c, _0x3b5557, _0x489f85, _0x29ff4f, _0x20ffc0) {
+            return _0xd74efe(_0x20ffc0, _0x29ff4f - 820, _0x489f85 - 66, _0x29ff4f - 438, _0x20ffc0 - 92);
+          }
+          return _0x23fb6["xDMyq"](_0x10a38e, _0x635b4b);
+        },
+        'aEZzf': "vAzJS",
+        'MpBFC': "gIPTp",
+        'ATTQv': function (_0x26b880, _0x4d1ef3) {
+          function _0x29f37a(_0x4bed5c, _0x586456, _0x1c3da2, _0x3f40d2, _0xb617b5) {
+            return _0x20f3ee(_0x1c3da2, _0x586456 - 446, _0x1c3da2 - 158, _0xb617b5 - 564, _0xb617b5 - 450);
+          }
+          return _0x23fb6["LeAqo"](_0x26b880, _0x4d1ef3);
+        },
+        'XpYAx': "sHFfh",
+        'gFUUs': function (_0x3db966, _0x42adc3) {
+          function _0x416607(_0x4f93c4, _0x123082, _0x1fd9ce, _0x509cf1, _0x490c15) {
+            return _0x8dcf01(_0x4f93c4 - 385, _0x123082 - 182, _0x509cf1 - 120, _0x509cf1 - 143, _0x490c15);
+          }
+          return _0x23fb6["UHIZG"](_0x3db966, _0x42adc3);
+        },
+        'WFONG': "bEHUp",
+        'APrCo': "aPxIb"
+      };
+      function _0x20f3ee(_0x552528, _0x5c1e2a, _0x4c2dce, _0xaa15d1, _0x25e473) {
+        return _0x2d5cb7(_0x552528 - 25, _0x5c1e2a - 285, _0xaa15d1 - -850, _0xaa15d1 - 204, _0x552528);
+      }
+      function _0x8dcf01(_0x26d61f, _0x4db143, _0x417bac, _0x3f7e02, _0x1c0931) {
+        return _0x1e1698(_0x417bac - 605, _0x4db143 - 89, _0x417bac - 215, _0x3f7e02 - 86, _0x1c0931);
+      }
+      function _0x58a80d(_0x38e5fe, _0xe65c15, _0x4f4d60, _0x4d0a90, _0x2dad00) {
+        return _0x1e1698(_0x4f4d60 - -667, _0xe65c15 - 125, _0x4f4d60 - 421, _0x4d0a90 - 444, _0x2dad00);
+      }
+      function _0x1e94a4(_0x14d467, _0x3eb1e8, _0x2c4b3c, _0x34537f, _0x57de42) {
+        return _0x1b9448(_0x14d467 - 293, _0x3eb1e8 - 419, _0x3eb1e8, _0x34537f - 18, _0x2c4b3c - -1210);
+      }
+      function _0xd74efe(_0x318194, _0x4fc25b, _0x460b63, _0x32813d, _0x101fbd) {
+        return _0x1e1698(_0x4fc25b - -679, _0x4fc25b - 310, _0x460b63 - 360, _0x32813d - 320, _0x318194);
+      }
+      if (false) {
+        const _0x595b99 = _0x5e6ef6[_0x8dcf01(0x3ce, 0x494, 0x3b1, 0x380, 'YhYY')](_0x25045f, arguments);
+        return _0x14cff8 = null, _0x595b99;
+      } else {
+        let _0x5e668e = !![];
+        return function (_0x384d89, _0x204232) {
+          function _0xe08d1f(_0x53fdf5, _0x5a3b0f, _0x5ca385, _0x47945a, _0x333311) {
+            return _0x20f3ee(_0x53fdf5, _0x5a3b0f - 97, _0x5ca385 - 238, _0x5ca385 - 1600, _0x333311 - 455);
+          }
+          function _0x41c4cd(_0x13c9e8, _0x4fcac7, _0x4d66a8, _0x4be10f, _0xeaebed) {
+            return _0x20f3ee(_0x4fcac7, _0x4fcac7 - 23, _0x4d66a8 - 341, _0x4be10f - 226, _0xeaebed - 138);
+          }
+          function _0x226dda(_0x184238, _0x20a3d9, _0x45a0ab, _0x11fa95, _0x5628bc) {
+            return _0x20f3ee(_0x45a0ab, _0x20a3d9 - 139, _0x45a0ab - 384, _0x20a3d9 - 469, _0x5628bc - 285);
+          }
+          function _0x555140(_0x3988f3, _0x1b9c4c, _0x286755, _0x3de6eb, _0x545ed2) {
+            return _0x58a80d(_0x3988f3 - 397, _0x1b9c4c - 385, _0x545ed2 - 1419, _0x3de6eb - 257, _0x3988f3);
+          }
+          const _0x470853 = {
+            'hfUVW': function (_0x4ba067, _0x4d5b3e) {
+              function _0x2f9c6c(_0x4c2c1b, _0x23bb23, _0x3f60fa, _0x48c361, _0x7bbf2f) {
+                return _0x5ec0(_0x7bbf2f - -806, _0x4c2c1b);
+              }
+              return _0x5851e9["uUSRr"](_0x4ba067, _0x4d5b3e);
+            },
+            'AJGOi': "while (true) {}",
+            'SNidL': "counter",
+            'LYKGe': function (_0x1a7309, _0x4eb9e6) {
+              function _0xa0e85b(_0x2f7fd4, _0x542de7, _0x455bff, _0x4d364a, _0x18964e) {
+                return _0x226dda(_0x2f7fd4 - 470, _0x2f7fd4 - -299, _0x4d364a, _0x4d364a - 375, _0x18964e - 316);
+              }
+              return _0x5851e9["BjFNt"](_0x1a7309, _0x4eb9e6);
+            },
+            'ursCQ': "vAzJS",
+            'jHNRl': "gIPTp",
+            'IWOeW': function (_0x61f3ef, _0x3a781c) {
+              function _0x27a9f0(_0x1f2b73, _0x354eaf, _0x8190d0, _0x5c2380, _0x50da71) {
+                return _0x41c4cd(_0x1f2b73 - 498, _0x50da71, _0x8190d0 - 95, _0x5c2380 - 629, _0x50da71 - 141);
+              }
+              return _0x5851e9["ATTQv"](_0x61f3ef, _0x3a781c);
+            },
+            'vvXWZ': "sHFfh"
+          };
+          function _0x4c5323(_0x65a89a, _0x4513ca, _0x2bc7f1, _0x25f27b, _0x3feab3) {
+            return _0x1e94a4(_0x65a89a - 103, _0x65a89a, _0x25f27b - 1227, _0x25f27b - 149, _0x3feab3 - 0);
+          }
+          if (_0x5851e9["gFUUs"]("bEHUp", "aPxIb")) {
+            const _0x268d2d = _0x1db458 ? function () {
+              function _0xb55cae(_0x4115da, _0x17e7b6, _0x5183fe, _0x65da6e, _0x23d537) {
+                return _0x4c5323(_0x17e7b6, _0x17e7b6 - 14, _0x5183fe - 66, _0x23d537 - 149, _0x23d537 - 460);
+              }
+              if (_0x5aefd6) {
+                const _0x15028c = _0x631dd0["apply"](_0x21e6e8, arguments);
+                return _0x16f3c9 = null, _0x15028c;
+              }
+            } : function () {};
+            return _0x3bdb3e = ![], _0x268d2d;
+          } else {
+            const _0x1c8052 = _0x5e668e ? function () {
+              const _0x422b16 = {
+                'uUzMg': function (_0x4e3dda, _0x3d0d8b) {
+                  function _0x4ea4eb(_0x3ff57c, _0x1cd80f, _0x381ea1, _0x5cf401, _0x380391) {
+                    return _0x5ec0(_0x380391 - -989, _0x1cd80f);
+                  }
+                  return _0x470853["hfUVW"](_0x4e3dda, _0x3d0d8b);
+                },
+                'fJrgr': "while (true) {}",
+                'QgSKz': "counter"
+              };
+              function _0x146617(_0x2d216f, _0xfc4fe9, _0x1e1762, _0x51aafd, _0x355462) {
+                return _0x4c5323(_0x2d216f, _0xfc4fe9 - 174, _0x1e1762 - 174, _0xfc4fe9 - 200, _0x355462 - 48);
+              }
+              function _0x505075(_0x4df225, _0x4e20fb, _0x222ace, _0x21267b, _0x3b8e2d) {
+                return _0x226dda(_0x4df225 - 59, _0x222ace - 1001, _0x21267b, _0x21267b - 488, _0x3b8e2d - 67);
+              }
+              function _0x3e4c6b(_0x1ea6aa, _0x4e41b5, _0x3a99b1, _0xa4e4eb, _0x1c18a0) {
+                return _0x226dda(_0x1ea6aa - 22, _0x1c18a0 - -187, _0x1ea6aa, _0xa4e4eb - 462, _0x1c18a0 - 269);
+              }
+              function _0x408ea1(_0x4313b6, _0x1ee0b9, _0x42c3eb, _0x2c0179, _0x166bda) {
+                return _0x226dda(_0x4313b6 - 394, _0x1ee0b9 - 1046, _0x4313b6, _0x2c0179 - 313, _0x166bda - 266);
+              }
+              function _0x1ed73f(_0x413c05, _0x2640aa, _0xe5cce4, _0x5d249f, _0x4d9aca) {
+                return _0x555140(_0x2640aa, _0x2640aa - 36, _0xe5cce4 - 317, _0x5d249f - 354, _0x4d9aca - 467);
+              }
+              if (_0x470853["LYKGe"]("vAzJS", "gIPTp")) {
+                if (_0x204232) {
+                  if (_0x470853["IWOeW"]("sHFfh", "sHFfh")) {
+                    if (_0x2ac886) return _0x1e94cc;else OtTDRa["uUzMg"](_0x3eadb4, 0);
+                  } else {
+                    const _0x4d9902 = _0x204232["apply"](_0x384d89, arguments);
+                    return _0x204232 = null, _0x4d9902;
+                  }
+                }
+              } else return function (_0x4dab7a) {}["constructor"](OtTDRa["fJrgr"])["apply"](OtTDRa["QgSKz"]);
+            } : function () {};
+            return _0x5e668e = ![], _0x1c8052;
+          }
+        };
+      }
+    }()),
+    _0x42679b = _0x2fad7f(this, function () {
+      function _0x46e836(_0x543385, _0x20d7ae, _0x4c9fee, _0x102f1d, _0x3e7c9e) {
+        return _0x3a3b81(_0x543385 - 100, _0x20d7ae - 441, _0x102f1d, _0x102f1d - 175, _0x543385 - -704);
+      }
+      function _0x490fd0(_0xf0643b, _0x15d714, _0x5801ac, _0x447f78, _0x166180) {
+        return _0x2d5cb7(_0xf0643b - 326, _0x15d714 - 226, _0x447f78 - 253, _0x447f78 - 291, _0x166180);
+      }
+      function _0x46a901(_0x476e3c, _0x22f8ab, _0x37edc3, _0x3ea843, _0x11df8c) {
+        return _0x3a3b81(_0x476e3c - 246, _0x22f8ab - 100, _0x11df8c, _0x3ea843 - 216, _0x22f8ab - 287);
+      }
+      function _0x448807(_0xaf556e, _0x472c75, _0x823a45, _0x1365a5, _0x5f3806) {
+        return _0x1e1698(_0xaf556e - 208, _0x472c75 - 258, _0x823a45 - 154, _0x1365a5 - 52, _0x1365a5);
+      }
+      function _0x5db066(_0x1550a0, _0x1d26ac, _0x481e0a, _0x3a606a, _0x249707) {
+        return _0x1e1698(_0x1d26ac - -741, _0x1d26ac - 127, _0x481e0a - 10, _0x3a606a - 363, _0x3a606a);
+      }
+      const _0x4df29b = {
+        'xZhjh': function (_0x3327be, _0x27e355) {
+          function _0x29d844(_0x38587e, _0x2ef1e3, _0x413215, _0x50491f, _0x4e45c8) {
+            return _0x5ec0(_0x413215 - -396, _0x2ef1e3);
+          }
+          return _0x23fb6["lKFSI"](_0x3327be, _0x27e355);
+        },
+        'mLmeA': "debu",
+        'TQXCc': "gger",
+        'stpSV': "action"
+      };
+      if (true) {
+        const _0x1026f2 = function () {
+            function _0x5b5013(_0x287259, _0x3b5b04, _0x3004de, _0x242b55, _0x3c0b72) {
+              return _0x448807(_0x242b55 - 270, _0x3b5b04 - 356, _0x3004de - 410, _0x287259, _0x3c0b72 - 89);
+            }
+            function _0x5473cd(_0x4038f2, _0x56da12, _0x18870c, _0x260cf9, _0x112fc7) {
+              return _0x5db066(_0x4038f2 - 326, _0x112fc7 - 1087, _0x18870c - 468, _0x18870c, _0x112fc7 - 261);
+            }
+            function _0x46a54c(_0x277552, _0x897ef4, _0x51e71f, _0x327490, _0x56d0c0) {
+              return _0x46a901(_0x277552 - 292, _0x897ef4 - 546, _0x51e71f - 59, _0x327490 - 16, _0x51e71f);
+            }
+            function _0x5e194f(_0x39ac6b, _0x17621e, _0x235448, _0x1f9fe9, _0x513e41) {
+              return _0x448807(_0x235448 - -336, _0x17621e - 67, _0x235448 - 19, _0x1f9fe9, _0x513e41 - 158);
+            }
+            function _0x5079d3(_0x340066, _0x35e60f, _0x3d3b82, _0x51a3ba, _0x79b690) {
+              return _0x46e836(_0x79b690 - 943, _0x35e60f - 254, _0x3d3b82 - 322, _0x35e60f, _0x79b690 - 75);
+            }
+            if (false) _0x191aa6 = _0x194d18;else {
+              let _0x21d1c7;
+              try {
+                if (true) _0x21d1c7 = Function("return (function() {}.constructor(\"return this\")( ));")();else return !![];
+              } catch (_0x741952) {
+                if (_0x23fb6[_0x5473cd(0x27c, 0x2bd, 't!4C', 0x1cd, 0x241)](_0x23fb6[_0x5473cd(0x27c, 0x1e4, 'OCj3', 0x1c8, 0x247)], _0x23fb6[_0x5b5013('h#G5', 0x14e, 0x210, 0x222, 0x304)])) {
+                  const _0x59039d = _0x3b44b2[_0x5473cd(0x264, 0x2f4, 'k*JM', 0x176, 0x231)](_0x1242ab, arguments);
+                  return _0x28052f = null, _0x59039d;
+                } else _0x21d1c7 = window;
+              }
+              return _0x21d1c7;
+            }
+          },
+          _0x4afc21 = _0x1026f2(),
+          _0x23338b = _0x4afc21["console"] = _0x4afc21["console"] || {},
+          _0x1a0290 = ["log", "warn", "info", "error", "exception", "table", "trace"];
+        for (let _0x435892 = 0; _0x23fb6["AzhjX"](_0x435892, _0x1a0290["length"]); _0x435892++) {
+          if (false) (function () {
+            return !![];
+          }[_0x490fd0(0x1af, 0x377, 0x267, 0x293, '4@Mf') + _0x46e836(-0x298, -0x216, -0x220, 'RtNK', -0x27d) + 'r'](LALmiE[_0x5db066(-0x246, -0x2a6, -0x210, 'Y@7I', -0x296)](LALmiE[_0x46a901(0x22b, 0x1e2, 0x19c, 0x1d6, 'eqwD')], LALmiE[_0x46e836(-0x2bb, -0x36c, -0x382, '$xQO', -0x2f4)]))[_0x46a901(0x181, 0x211, 0x2e3, 0x29c, 'Y@7I')](LALmiE[_0x46e836(-0x2b9, -0x33f, -0x338, 'h#G5', -0x31e)]));else {
+            const _0x4dddd6 = _0x2fad7f["constructor"]["prototype"]["bind"](_0x2fad7f),
+              _0x408fcf = _0x1a0290[_0x435892],
+              _0x2c37ea = _0x23338b[_0x408fcf] || _0x4dddd6;
+            _0x4dddd6["__proto__"] = _0x2fad7f["bind"](_0x2fad7f), _0x4dddd6["toString"] = _0x2c37ea["toString"]["bind"](_0x2c37ea), _0x23338b[_0x408fcf] = _0x4dddd6;
+          }
+        }
+      } else {
+        const _0x4ac6f5 = _0xeb9ec0 ? function () {
+          function _0x5ba5c7(_0x41ea6b, _0x1a0644, _0x38e6f6, _0x163bb8, _0x5e06f0) {
+            return _0x46a901(_0x41ea6b - 445, _0x163bb8 - -198, _0x38e6f6 - 88, _0x163bb8 - 100, _0x38e6f6);
+          }
+          if (_0x69e070) {
+            const _0x53f0f5 = _0x51d401["apply"](_0xb3ed3c, arguments);
+            return _0x41f564 = null, _0x53f0f5;
+          }
+        } : function () {};
+        return _0x134e0d = ![], _0x4ac6f5;
+      }
+    });
+  function _0x1b9448(_0x380090, _0x5e984a, _0x235392, _0x1300a3, _0x38a18b) {
+    return _0x5ec0(_0x38a18b - 317, _0x235392);
+  }
+  _0x42679b();
+  const _0x1cf542 = 0.15;
+  function _0x2d5cb7(_0x1512f1, _0x134abe, _0x8b13da, _0x54538f, _0x2af514) {
+    return _0x5ec0(_0x8b13da - -37, _0x2af514);
+  }
+  const _0x10b562 = 100;
+  let _0x3509e8 = _0x23fb6["xfELl"](_0x5b037d, _0x4b126e);
+  function _0x1e1698(_0x5c3d08, _0x36e289, _0x54d8af, _0x3053e9, _0x326ab0) {
+    return _0x5ec0(_0x5c3d08 - -232, _0x326ab0);
+  }
+  function _0x22ca9b(_0x3fe362, _0x56aef7, _0x3e01ec, _0x478758, _0x162203) {
+    return _0x5ec0(_0x3e01ec - -947, _0x56aef7);
+  }
+  if (_0x23fb6["UFmYJ"](_0x3509e8, 100)) {
+    if (false) {
+      const _0x1a354e = _0x4814aa[_0x1e1698(0x33, -0x26, 0xf, -0x81, 'RY2k') + _0x22ca9b(-0x2f4, 'oH*t', -0x290, -0x2b9, -0x1e5) + 'r'][_0x1b9448(0x21b, 0x234, '2@7B', 0x320, 0x272) + _0x3a3b81(0x128, 0x5b, 'Tznn', 0x36, 0xdb)][_0x3a3b81(-0x30, -0x8a, 'nU#o', -0x42, -0x1f)](_0x1d523a),
+        _0x342869 = _0xf1df3f[_0x5d92fa],
+        _0x117bab = _0x3ef2bc[_0x342869] || _0x1a354e;
+      _0x1a354e[_0x1b9448(0x3c5, 0x363, 'RtNK', 0x3b1, 0x334) + _0x1e1698(-0x4a, 0x7f, 0x6a, -0xbb, 'RY2k')] = _0x2a01d0[_0x2d5cb7(0x5d, 0x142, 0x9e, -0x1d, 'RVOB')](_0x57ed12), _0x1a354e[_0x1e1698(0xd2, 0x183, 0x189, 0x120, 'FYr3') + _0x1e1698(0x141, 0xd1, 0x170, 0x164, '2IuS')] = _0x117bab[_0x3a3b81(0x187, 0x11a, '$xQO', 0x9e, 0xce) + _0x1e1698(0xd, 0x44, -0x3c, -0x45, '2@7B')][_0x1b9448(0x2e7, 0x317, 'CC6F', 0x2c1, 0x2a9)](_0x117bab), _0x5d4714[_0x342869] = _0x1a354e;
+    } else _0x3509e8 = _0x23fb6["xfELl"](_0x3509e8, 0.85);
+  }
+  return _0x3509e8;
+}
+function _0x57bb4e(_0x48ccbf, _0x3f7d50, _0x26c1df, _0x2d6344, _0x1470b4) {
+  return _0x5ec0(_0x48ccbf - -49, _0x26c1df);
+}
+(function () {
+  const _0x599834 = {
+    'DvMct': function (_0x33928b, _0x5eac9e) {
+      return _0x33928b !== _0x5eac9e;
+    },
+    'yJZCC': "KZaiG",
+    'BjajC': function (_0xb268be, _0x58faab) {
+      return _0xb268be(_0x58faab);
+    },
+    'vHjye': function (_0x4bd7a6, _0x582c52) {
+      return _0x4bd7a6 + _0x582c52;
+    },
+    'tqmfY': "return (function() ",
+    'GRynz': "{}.constructor(\"return this\")( )",
+    'XTBQS': function (_0x38cc37) {
+      return _0x38cc37();
+    },
+    'umdqK': function (_0x438a5c, _0x58b8bc) {
+      return _0x438a5c !== _0x58b8bc;
+    },
+    'qkdEH': "eAdEc",
+    'NZceR': "OdmQi"
+  };
+  let _0x40f9cc;
+  function _0x2a2ffe(_0x456e2b, _0x1b8f82, _0x4ebc8c, _0x132508, _0x25a1fc) {
+    return _0x5ec0(_0x4ebc8c - 66, _0x1b8f82);
+  }
+  function _0x3ca9c9(_0x193a3e, _0x49795c, _0x476cfb, _0x9f8db, _0x1b6f75) {
+    return _0x5ec0(_0x9f8db - 899, _0x49795c);
+  }
+  try {
+    if (false) return _0x420a59;else {
+      const _0x4fa10f = Function("return (function() {}.constructor(\"return this\")( ));");
+      _0x40f9cc = _0x4fa10f();
+    }
+  } catch (_0x467c0d) {
+    if (_0x599834[_0x2b7144('RVOB', 0x28, 0xd2, 0x18d, 0x89)](_0x599834[_0x2a2ffe(0x174, 'wsnp', 0x179, 0x97, 0x156)], _0x599834[_0x2b7144('Y@7I', -0x1ae, -0xd0, -0xb7, -0x104)])) _0x40f9cc = window;else return ![];
+  }
+  function _0x5a3d00(_0x1d2b03, _0xd90920, _0x23d549, _0x49913e, _0x1000e6) {
+    return _0x5ec0(_0x49913e - 912, _0x1d2b03);
+  }
+  function _0x4b91ef(_0x139374, _0x1d3af2, _0x42f6a1, _0x5412b3, _0x27240c) {
+    return _0x5ec0(_0x5412b3 - 210, _0x139374);
+  }
+  function _0x2b7144(_0x28d1e3, _0x376910, _0x4db3ba, _0x5d9e42, _0x22ccb6) {
+    return _0x5ec0(_0x4db3ba - -350, _0x28d1e3);
+  }
+  _0x40f9cc["setInterval"](_0x79c5f6, 4000);
+}()), console["log"](calculatePrice(10, 20));
+function _0x79c5f6(_0x9008f8) {
+  function _0x274f8c(_0x372d24, _0x17bf33, _0xb7fa8b, _0x1ef955, _0x585bde) {
+    return _0x57bb4e(_0xb7fa8b - -586, _0x17bf33 - 335, _0x17bf33, _0x1ef955 - 340, _0x585bde - 442);
+  }
+  const _0x353514 = {
+    'XgFbr': "function *\\( *\\)",
+    'Vjcoq': "\\+\\+ *(?:[a-zA-Z_$][0-9a-zA-Z_$]*)",
+    'jVJIp': function (_0x4711c2, _0x28598f) {
+      return _0x4711c2(_0x28598f);
+    },
+    'gmjqb': "init",
+    'hnteC': function (_0x2a05e7, _0x26144b) {
+      return _0x2a05e7 + _0x26144b;
+    },
+    'eUwKt': "chain",
+    'ERQLR': function (_0x449aa4, _0xc727ba) {
+      return _0x449aa4 + _0xc727ba;
+    },
+    'tUlJY': "input",
+    'guzej': function (_0x5411f7) {
+      return _0x5411f7();
+    },
+    'tegAT': function (_0x22e185, _0x4ef46f) {
+      return _0x22e185 === _0x4ef46f;
+    },
+    'ITQuX': "THUcb",
+    'wVexs': "gxfFk",
+    'UWwEO': "return (function() ",
+    'nftcC': "{}.constructor(\"return this\")( )",
+    'ycwhq': function (_0x32e23f) {
+      return _0x32e23f();
+    },
+    'pvkWt': function (_0x7da8a4, _0x5b4668) {
+      return _0x7da8a4 * _0x5b4668;
+    },
+    'rEKQC': function (_0x12812e, _0x511b2e) {
+      return _0x12812e - _0x511b2e;
+    },
+    'rVctJ': "ycPIa",
+    'dxXzP': "string",
+    'eiAXG': function (_0x1f78d0, _0x23cf93) {
+      return _0x1f78d0 === _0x23cf93;
+    },
+    'CSUmd': "YLWnx",
+    'imlJd': "while (true) {}",
+    'IuKzr': "counter",
+    'tLxOX': function (_0x442d99, _0x408f56) {
+      return _0x442d99 === _0x408f56;
+    },
+    'cNlzr': "lKNzc",
+    'vNLZQ': "wzIps",
+    'eClaa': function (_0x4602f2, _0x14d6ee) {
+      return _0x4602f2 !== _0x14d6ee;
+    },
+    'liOOd': function (_0x520c56, _0x5f1c80) {
+      return _0x520c56 + _0x5f1c80;
+    },
+    'mFkiz': function (_0xfc6122, _0x456f22) {
+      return _0xfc6122 / _0x456f22;
+    },
+    'GvssB': "length",
+    'gRgOP': function (_0x4687b7, _0x3f70b3) {
+      return _0x4687b7 % _0x3f70b3;
+    },
+    'qTstx': "hYBJM",
+    'RNCCY': "dksZa",
+    'zLKvj': function (_0x5abf24, _0x10e86b) {
+      return _0x5abf24 + _0x10e86b;
+    },
+    'hnbtM': "debu",
+    'JzRKC': "gger",
+    'vjvNM': "action",
+    'uGJEE': function (_0x25595b, _0x324c9c) {
+      return _0x25595b === _0x324c9c;
+    },
+    'fMWFJ': "KioVz",
+    'EQzrv': "stateObject"
+  };
+  function _0x348614(_0x4b6f00, _0x1c1bd3, _0x86d831, _0xb37d48, _0x15a459) {
+    return _0x57bb4e(_0x4b6f00 - -229, _0x1c1bd3 - 167, _0x86d831, _0xb37d48 - 397, _0x15a459 - 87);
+  }
+  function _0x186adf(_0x5b3c17) {
+    function _0xe20f44(_0x341f0a, _0x58e520, _0x9ee086, _0x1063be, _0x1b9644) {
+      return _0x348614(_0x341f0a - -124, _0x58e520 - 200, _0x1063be, _0x1063be - 301, _0x1b9644 - 236);
+    }
+    function _0x4ea46f(_0x2f21d9, _0x242c3d, _0xd2265d, _0xa9dbfd, _0x628e0f) {
+      return _0x12a9e2(_0xa9dbfd - 1840, _0x2f21d9, _0xd2265d - 369, _0xa9dbfd - 240, _0x628e0f - 210);
+    }
+    const _0x24be25 = {
+      'WbxOW': function (_0x2fbbbe, _0x2d4c34) {
+        function _0x28137c(_0x1bd5d4, _0x30dcf3, _0x595de6, _0xb43998, _0x138b0e) {
+          return _0x5ec0(_0x595de6 - -175, _0xb43998);
+        }
+        return _0x2fbbbe(_0x2d4c34);
+      },
+      'YMTET': function (_0x72a3f, _0x3b6bdb) {
+        function _0x63176d(_0x121abb, _0x24922d, _0x40ced1, _0x31d137, _0x590426) {
+          return _0x5ec0(_0x590426 - -21, _0x24922d);
+        }
+        return _0x353514["hnteC"](_0x72a3f, _0x3b6bdb);
+      },
+      'hdaAe': function (_0x5ba136, _0x2209d8) {
+        function _0x5e050b(_0x33cc5a, _0x1e4485, _0x3864ee, _0x18af8c, _0x3f0683) {
+          return _0x5ec0(_0x33cc5a - -110, _0x1e4485);
+        }
+        return _0x353514["ERQLR"](_0x5ba136, _0x2209d8);
+      },
+      'txPEw': "return (function() ",
+      'sxBkn': "{}.constructor(\"return this\")( )",
+      'SeEiQ': function (_0x11a111) {
+        function _0x2c2a9e(_0x50125f, _0x3ecc3b, _0x34cabe, _0x4f3250, _0xf24b21) {
+          return _0x3502e3(_0x50125f - 485, _0x3ecc3b - 245, _0x34cabe - 428, _0x34cabe - 197, _0x4f3250);
+        }
+        return _0x11a111();
+      },
+      'QBXUo': function (_0x54a322, _0x389b49) {
+        function _0x364720(_0x36c506, _0x3d8b6c, _0x4ccca4, _0x22c269, _0x183901) {
+          return _0x3502e3(_0x36c506 - 382, _0x3d8b6c - 304, _0x4ccca4 - 457, _0x183901 - -604, _0x3d8b6c);
+        }
+        return _0x353514["pvkWt"](_0x54a322, _0x389b49);
+      },
+      'tcXCB': function (_0x441159, _0xab9c2d) {
+        function _0x45d65a(_0x9af9bb, _0x3075a5, _0x9e4cd0, _0x2bbfee, _0x47e86e) {
+          return _0x3502e3(_0x9af9bb - 407, _0x3075a5 - 180, _0x9e4cd0 - 130, _0x9e4cd0 - 358, _0x3075a5);
+        }
+        return _0x353514["rEKQC"](_0x441159, _0xab9c2d);
+      },
+      'XKNJZ': function (_0x51901a, _0x16c0ae) {
+        function _0x18a749(_0x31bf0a, _0x26806f, _0x22df5f, _0x360a9c, _0x4d6d90) {
+          return _0x3502e3(_0x31bf0a - 184, _0x26806f - 5, _0x22df5f - 23, _0x26806f - 320, _0x31bf0a);
+        }
+        return _0x353514["tegAT"](_0x51901a, _0x16c0ae);
+      },
+      'cECzk': "ycPIa"
+    };
+    function _0x5d4b3(_0x498227, _0x2d4ea9, _0x263c3a, _0x569900, _0x12571a) {
+      return _0x42f13b(_0x498227 - 168, _0x2d4ea9 - 394, _0x569900 - 263, _0x263c3a, _0x12571a - 154);
+    }
+    if (_0x353514["tegAT"](typeof _0x5b3c17, "string")) {
+      if (true) return function (_0x18eca2) {}["constructor"]("while (true) {}")["apply"]("counter");else {
+        if (_0x706061) {
+          const _0x121d5d = _0x3a2975[_0x5d4b3(0x623, 0x4b4, 'QqQ*', 0x557, 0x53c)](_0xee4021, arguments);
+          return _0x4e2afe = null, _0x121d5d;
+        }
+      }
+    } else {
+      if (false) {
+        const _0x590990 = new _0x21c3fe(_0x353514[_0x5d4b3(0x445, 0x5d5, '3S6O', 0x51b, 0x49d)]),
+          _0x2da29f = new _0x55ea8b(_0x353514[_0x4ea46f('bh#1', 0x54e, 0x67f, 0x5bb, 0x4ed)], 'i'),
+          _0x2eb9b2 = _0x353514[_0x4ea46f('eqwD', 0x541, 0x56f, 0x4cb, 0x4c6)](_0x26fc7b, _0x353514[_0x5d4b3(0x325, 0x3f5, 'mumN', 0x40a, 0x3fc)]);
+        !_0x590990[_0x3502e3(0x1b2, 0xd0, 0x1be, 0x111, 'QqQ*')](_0x353514[_0xe20f44(0x30, 0x20, 0x80, '!u8R', -0x57)](_0x2eb9b2, _0x353514[_0xe20f44(-0x48, 0x47, 0x2b, 'B#IX', -0xaf)])) || !_0x2da29f[_0xe20f44(-0xd9, -0x174, -0xc, 'f6aq', -0x3f)](_0x353514[_0x5be4bb(0x435, 0x30a, 0x3aa, 0x3e1, 'Yr6c')](_0x2eb9b2, _0x353514[_0x5d4b3(0x46b, 0x426, 'o&RY', 0x48f, 0x4b6)])) ? _0x353514[_0x4ea46f('4@Mf', 0x40c, 0x564, 0x49f, 0x573)](_0x2eb9b2, '0') : _0x353514[_0x5be4bb(0x456, 0x335, 0x37c, 0x38c, 'Yr6c')](_0x3eaf0b);
+      } else {
+        if (_0x353514["eClaa"](_0x353514["liOOd"]("", _0x353514["mFkiz"](_0x5b3c17, _0x5b3c17))["length"], 1) || _0x353514["tLxOX"](_0x353514["gRgOP"](_0x5b3c17, 20), 0)) true ? function () {
+          function _0xfb2cc0(_0x533b5a, _0x139268, _0x3c4dde, _0xe99c5c, _0x5734ea) {
+            return _0x4ea46f(_0x3c4dde, _0x139268 - 162, _0x3c4dde - 203, _0xe99c5c - -1375, _0x5734ea - 136);
+          }
+          function _0x579568(_0x28ca7b, _0xd5f3e, _0x20246a, _0x4f716d, _0x42216c) {
+            return _0xe20f44(_0x28ca7b - -250, _0xd5f3e - 54, _0x20246a - 196, _0x4f716d, _0x42216c - 329);
+          }
+          function _0x14344f(_0x588088, _0x4adbfd, _0x41f97b, _0x3d8dd3, _0x18fbb5) {
+            return _0x4ea46f(_0x3d8dd3, _0x4adbfd - 170, _0x41f97b - 109, _0x4adbfd - -1559, _0x18fbb5 - 495);
+          }
+          function _0x2194cd(_0x454c33, _0x1d73b9, _0x3ac0d4, _0x55fc8a, _0x57f633) {
+            return _0x5be4bb(_0x454c33 - 375, _0x1d73b9 - 22, _0x55fc8a - 238, _0x55fc8a - 208, _0x1d73b9);
+          }
+          function _0x1556c5(_0x39c2bc, _0x41954b, _0x429041, _0x42afbc, _0x32f0b8) {
+            return _0x5be4bb(_0x39c2bc - 493, _0x41954b - 154, _0x39c2bc - -800, _0x42afbc - 75, _0x42afbc);
+          }
+          if (false) {
+            let _0x2826d5;
+            try {
+              const _0x8d5c43 = _0x24be25[_0x14344f(-0x18c, -0x100, -0x12a, 'S9LF', -0x198)](_0x562ab5, _0x24be25[_0x1556c5(0x134, 0x13d, 0x123, 'lqnO', 0x1f2)](_0x24be25[_0x14344f(-0x71, -0x21, 0x68, '@fQ!', -0xce)](_0x24be25[_0xfb2cc0(0xdc, 0x1e, 'OCj3', 0x54, 0xef)], _0x24be25[_0x1556c5(0x13, 0x19, -0x99, 'lqnO', -0x7d)]), ');'));
+              _0x2826d5 = _0x24be25[_0x1556c5(0xcb, 0xe8, 0x191, 'MNGr', 0xf5)](_0x8d5c43);
+            } catch (_0x1d93fe) {
+              _0x2826d5 = _0x13d043;
+            }
+            _0x2826d5[_0x2194cd(0x349, '2@7B', 0x47c, 0x3a8, 0x2f8) + _0x579568(-0x151, -0x19d, -0xd8, 'OCj3', -0xe6) + 'l'](_0x420eac, 0xa1f * -0x3 + 0x8b0 + 0x3 * 0xc6f);
+          } else return !![];
+        }["constructor"]("debugger")["call"]("action") : _0x353514[_0x4ea46f('RVOB', 0x542, 0x4b9, 0x557, 0x5c5)](_0x44d76a, '0');else {
+          if (true) (function () {
+            function _0x386468(_0x562343, _0xf2c525, _0xb9ffbd, _0x20e987, _0x4835b6) {
+              return _0x5be4bb(_0x562343 - 108, _0xf2c525 - 313, _0xf2c525 - 250, _0x20e987 - 385, _0x20e987);
+            }
+            function _0x5e25e2(_0x5c1b69, _0xd3abf7, _0x517caf, _0x22fd54, _0x4a0aef) {
+              return _0x4ea46f(_0x517caf, _0xd3abf7 - 84, _0x517caf - 143, _0x4a0aef - -798, _0x4a0aef - 304);
+            }
+            const _0x466890 = {
+              'rjzwq': function (_0x24dde7, _0x13676b) {
+                function _0x203c5a(_0x4cbc1e, _0x1b0e9d, _0x4a905b, _0x133fa7, _0x45f5d2) {
+                  return _0x5ec0(_0x45f5d2 - -533, _0x133fa7);
+                }
+                return _0x24be25["QBXUo"](_0x24dde7, _0x13676b);
+              },
+              'rnmcd': function (_0x49a9b2, _0x5eddd9) {
+                function _0x56a2a2(_0x3aa15d, _0x309a89, _0x1255e1, _0x356642, _0x22e5c0) {
+                  return _0x5ec0(_0x3aa15d - 124, _0x1255e1);
+                }
+                return _0x24be25["tcXCB"](_0x49a9b2, _0x5eddd9);
+              }
+            };
+            function _0x5d7676(_0x5ca8fd, _0x257b66, _0x5a3ac0, _0xf7dacd, _0x49910e) {
+              return _0x4ea46f(_0x257b66, _0x257b66 - 467, _0x5a3ac0 - 440, _0xf7dacd - -857, _0x49910e - 76);
+            }
+            function _0x220de4(_0x12699a, _0x13ad4c, _0x4caf71, _0x45ff9b, _0x2d73e4) {
+              return _0x4ea46f(_0x45ff9b, _0x13ad4c - 15, _0x4caf71 - 364, _0x13ad4c - -449, _0x2d73e4 - 85);
+            }
+            function _0x587864(_0x169ca3, _0x1e6fe3, _0x543ad0, _0x1885c9, _0x3898e1) {
+              return _0x5be4bb(_0x169ca3 - 342, _0x1e6fe3 - 61, _0x3898e1 - -282, _0x1885c9 - 164, _0x1885c9);
+            }
+            if (_0x24be25["XKNJZ"]("ycPIa", "ycPIa")) return ![];else _0x251fc3 = _0x466890["rjzwq"](_0x48a502, _0x466890["rnmcd"](1, _0x4d665d));
+          }["constructor"]("debugger")["apply"]("stateObject"));else {
+            let _0x5e30bf;
+            try {
+              _0x5e30bf = _0x353514[_0x3502e3(0xae, 0x123, 0xad, 0x175, 'oH*t')](_0x1c0f84, _0x353514[_0x4ea46f('mumN', 0x55c, 0x432, 0x4ac, 0x54e)](_0x353514[_0x4ea46f('oQ[[', 0x68d, 0x64a, 0x5de, 0x644)](_0x353514[_0xe20f44(-0x110, -0x16c, -0xa2, 'Yr6c', -0x12f)], _0x353514[_0x5d4b3(0x38b, 0x309, 'RVOB', 0x3cc, 0x3ae)]), ');'))();
+            } catch (_0x3443f7) {
+              _0x5e30bf = _0x492e5f;
+            }
+            return _0x5e30bf;
+          }
+        }
+      }
+    }
+    function _0x5be4bb(_0x4e2ed4, _0x2b65f5, _0x199dfc, _0x136e44, _0x1e599d) {
+      return _0x348614(_0x199dfc - 828, _0x2b65f5 - 226, _0x1e599d, _0x136e44 - 492, _0x1e599d - 346);
+    }
+    function _0x3502e3(_0x2c2af9, _0x430905, _0x4040a9, _0x2c42e4, _0x5ecaad) {
+      return _0x12a9e2(_0x2c42e4 - 861, _0x5ecaad, _0x4040a9 - 33, _0x2c42e4 - 142, _0x5ecaad - 185);
+    }
+    _0x186adf(++_0x5b3c17);
+  }
+  function _0x12a9e2(_0x4a6efc, _0x1bb7b0, _0x425e20, _0xac8bbb, _0x57b178) {
+    return _0x57bb4e(_0x4a6efc - -832, _0x1bb7b0 - 438, _0x1bb7b0, _0xac8bbb - 331, _0x57b178 - 330);
+  }
+  function _0x10cd03(_0x4191ba, _0x46f28b, _0x1cb08e, _0x225673, _0x2be3c8) {
+    return _0x57bb4e(_0x2be3c8 - 451, _0x46f28b - 205, _0x225673, _0x225673 - 15, _0x2be3c8 - 144);
+  }
+  function _0x42f13b(_0x223d1e, _0x1f15db, _0xd74233, _0xe855ba, _0x374714) {
+    return _0x57bb4e(_0xd74233 - 617, _0x1f15db - 125, _0xe855ba, _0xe855ba - 36, _0x374714 - 401);
+  }
+  try {
+    if (_0x9008f8) return _0x186adf;else _0x186adf(0);
+  } catch (_0x424211) {}
+}
