+++ maldoca/js/ir/transforms/dead_code_elimination/tests/after_dynconstprop/run.generated.lit
@@ -1,0 +1,4 @@
+// RUN: CURRENT_FILE_BASENAME=$(basename %s .lit) && \
+// RUN: jsir_gen --input_file "$(dirname %s)"/input.js \
+// RUN:     --passes "source2ast,ast2hir,dead_code_elimination,hir2ast,ast2source" \
+// RUN:   | FileCheck --check-prefix SOURCE "$(dirname %s)"/output.generated.txt
