+++ maldoca/js/ir/transforms/dead_code_elimination/tests/unused_function/input.js
@@ -1,0 +1,18 @@
+function add(a, b) {
+    return a + b;
+}
+
+function unused_subtract(a, b) {
+    return a - b;
+}
+
+function multiply(a, b) {
+    return a * b;
+}
+
+function multiply_times_2(a, b) {
+    return multiply(a, b) * 2;
+}
+
+add(2, 3);
+multiply_times_2(10, 5);
