--- maldoca/js/ir/transforms/dead_code_elimination/tests/unused_functions/README.generated.md
@@ -1,7 +1,0 @@
-To run manually:
-
-```shell
-bazel run //maldoca/js/ir:jsir_gen -- \
-  --input_file $(pwd)/maldoca/js/ir/transforms/dead_code_elimination/tests/unused_functions/input.js \
-  --passes "source2ast,extract_prelude,erase_comments,ast2hir,dynconstprop,dead_code_elimination,hir2ast,ast2source"
-```
