+++ maldoca/js/ir/transforms/dead_code_elimination/tests/while_statement/input.js
@@ -1,7 +1,7 @@
 while(false){
-    console.log("To be eliminated");
+  console.log("To be eliminated");
 }
 
 while(true){
-    console.log("Should not be eliminated");   
+  console.log("Should not be eliminated");   
 }
