+++ maldoca/js/ir/transforms/dead_code_elimination/tests/while_statement/output.generated.txt
@@ -1,117 +1,3 @@
 // SOURCE:      while (true) {
 // SOURCE-NEXT:   console.log("Should not be eliminated");
 // SOURCE-NEXT: }
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: Prelude:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: ================================================================
-// SOURCE-EMPTY:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: Dataflow:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: jsir.file ({
-// SOURCE-NEXT:   ^bb0:
-// SOURCE-NEXT:       // <executable>
-// SOURCE-NEXT:       // State [default = <unknown>] { }
-// SOURCE-NEXT:     jsir.program {"script"} ({
-// SOURCE-NEXT:       ^bb0:
-// SOURCE-NEXT:           // <executable>
-// SOURCE-NEXT:           // State [default = <unknown>] { }
-// SOURCE-NEXT:         jshir.while_statement ({
-// SOURCE-NEXT:           ^bb0:
-// SOURCE-NEXT:               // <executable>
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:             %0 = jsir.boolean_literal {false}
-// SOURCE-NEXT:               // %0 = false
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:             jsir.expr_region_end (%0)
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:         }, {
-// SOURCE-NEXT:           ^bb0:
-// SOURCE-NEXT:               // <not executable>
-// SOURCE-NEXT:               // State [default = <uninitialized>] { }
-// SOURCE-NEXT:             jshir.block_statement ({
-// SOURCE-NEXT:               ^bb0:
-// SOURCE-NEXT:                   // <not executable>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:                 %0 = jsir.identifier {"console"}
-// SOURCE-NEXT:                   // %0 = <uninitialized>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:                 %1 = jsir.member_expression (%0) {#jsir<identifier   <L 2 C 12>, <L 2 C 15>, "log", 26, 29, 2, "log">}
-// SOURCE-NEXT:                   // %1 = <uninitialized>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:                 %2 = jsir.string_literal {#jsir<string_literal_extra "\22To be eliminated\22", "To be eliminated">, "To be eliminated"}
-// SOURCE-NEXT:                   // %2 = <uninitialized>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:                 %3 = jsir.call_expression (%1, %2)
-// SOURCE-NEXT:                   // %3 = <uninitialized>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:                 jsir.expression_statement (%3)
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:             }, {
-// SOURCE-NEXT:               ^bb0:
-// SOURCE-NEXT:                   // <not executable>
-// SOURCE-NEXT:                   // State [default = <uninitialized>] { }
-// SOURCE-NEXT:             })
-// SOURCE-NEXT:               // State [default = <uninitialized>] { }
-// SOURCE-NEXT:         })
-// SOURCE-NEXT:           // State [default = <unknown>] { }
-// SOURCE-NEXT:         jshir.while_statement ({
-// SOURCE-NEXT:           ^bb0:
-// SOURCE-NEXT:               // <executable>
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:             %0 = jsir.boolean_literal {true}
-// SOURCE-NEXT:               // %0 = true
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:             jsir.expr_region_end (%0)
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:         }, {
-// SOURCE-NEXT:           ^bb0:
-// SOURCE-NEXT:               // <executable>
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:             jshir.block_statement ({
-// SOURCE-NEXT:               ^bb0:
-// SOURCE-NEXT:                   // <executable>
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:                 %0 = jsir.identifier {"console"}
-// SOURCE-NEXT:                   // %0 = <unknown>
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:                 %1 = jsir.member_expression (%0) {#jsir<identifier   <L 6 C 12>, <L 6 C 15>, "log", 79, 82, 4, "log">}
-// SOURCE-NEXT:                   // %1 = <unknown>
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:                 %2 = jsir.string_literal {#jsir<string_literal_extra "\22Should not be eliminated\22", "Should not be eliminated">, "Should not be eliminated"}
-// SOURCE-NEXT:                   // %2 = "Should not be eliminated"
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:                 %3 = jsir.call_expression (%1, %2)
-// SOURCE-NEXT:                   // %3 = <unknown>
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:                 jsir.expression_statement (%3)
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:             }, {
-// SOURCE-NEXT:               ^bb0:
-// SOURCE-NEXT:                   // <executable>
-// SOURCE-NEXT:                   // State [default = <unknown>] { }
-// SOURCE-NEXT:             })
-// SOURCE-NEXT:               // State [default = <unknown>] { }
-// SOURCE-NEXT:         })
-// SOURCE-NEXT:           // State [default = <uninitialized>] { }
-// SOURCE-NEXT:     }, {
-// SOURCE-NEXT:       ^bb0:
-// SOURCE-NEXT:           // <executable>
-// SOURCE-NEXT:           // State [default = <unknown>] { }
-// SOURCE-NEXT:     })
-// SOURCE-NEXT:       // State [default = <unknown>] { }
-// SOURCE-NEXT: })
-// SOURCE-NEXT:   // State [default = <uninitialized>] { }
-// SOURCE-NEXT: ================================================================
-// SOURCE-EMPTY:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: Bindings:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: ================================================================
-// SOURCE-EMPTY:
-// SOURCE-NEXT: ================================================================
-// SOURCE-NEXT: Computed constants:
-// SOURCE-NEXT: ================================================================
-// SOURCE-EMPTY:
-// SOURCE-NEXT: ================================================================
