+++ maldoca/js/ir/transforms/dead_code_elimination/tests/while_statement/run.generated.lit
@@ -1,4 +1,4 @@
 // RUN: CURRENT_FILE_BASENAME=$(basename %s .lit) && \
 // RUN: jsir_gen --input_file "$(dirname %s)"/input.js \
-// RUN:     --passes "source2ast,extract_prelude,erase_comments,ast2hir,dynconstprop,dead_code_elimination,hir2ast,ast2source" \
+// RUN:     --passes "source2ast,ast2hir,dead_code_elimination,hir2ast,ast2source" \
 // RUN:   | FileCheck --check-prefix SOURCE "$(dirname %s)"/output.generated.txt
